# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital SaaS API",
    "description": "Public Swagger documentation for Hospital SaaS Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://YOUR_PRODUCTION_DOMAIN",
      "description": "Production (Replace manually)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/patients": {
      "get": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Search / List Patients",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "number" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Patients list" } }
      },
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Create Patient",
        "responses": { "200": { "description": "Patient created" } }
      }
    },
    "/api/patients/{id}": {
      "put": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Update Patient",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Patient updated" } }
      }
    },
    "/api/patients/export": {
      "get": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Export Patients CSV",
        "responses": { "200": { "description": "CSV file" } }
      }
    },
    "/api/patients/fetch-by-aadhaar": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Fetch Patient by Aadhaar",
        "responses": { "200": { "description": "Matched patients" } }
      }
    },
    "/api/patients/request-import-otp": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Request Import OTP",
        "responses": { "200": { "description": "OTP Sent" } }
      }
    },
    "/api/patients/verify-otp": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Verify OTP and Import Patient",
        "responses": { "200": { "description": "Patient imported" } }
      }
    },
...
  }
}